requirements.txt
python-dotenv
watchdog==6.0.0
Flask==3.1.1
requests
